BB2-2829 Create update access grant script.  Update Application.save to handle change to 13-month apps.

+++ apps/dot_ext/models.py
@@ -248,6 +248,8 @@ class Application(AbstractApplication):
                 app_from_db = None
                 try:
                     app_from_db = Application.objects.get(pk=self.id)
+                    self.copy_client_secret()
+                    super().save(*args, **kwargs)
                     if app_from_db is not None:
                         if self.data_access_type != app_from_db.data_access_type:
                             # log audit event: application data access type changed
@@ -257,26 +259,36 @@ class Application(AbstractApplication):
                                 "application_name": self.name,
                                 "data_access_type_old": app_from_db.data_access_type,
                                 "data_access_type_new": self.data_access_type,
-                                "grant_delete_start": datetime.now().strftime("%m/%d/%Y, %H:%M:%S"),
+                                "grant_start": datetime.now().strftime("%m/%d/%Y, %H:%M:%S"),
                             })
                             logger.info(log_dict)
-                            dag_deleted = DataAccessGrant.objects.filter(application=self).delete()
+                            if self.has_one_time_only_data_access():
+                                dag_deleted = DataAccessGrant.objects.filter(application=self).delete()
+                                end_time = time.time()
+                                delete_stats = {
+                                    "elapsed_seconds": end_time - start_time,
+                                    "number_of_grant_deleted": dag_deleted[0],
+                                    "grant_delete_complete": datetime.now().strftime("%m/%d/%Y, %H:%M:%S")
+                                }
+                                log_dict.update(delete_stats)
+                            elif "THIRTEEN_MONTH" in self.data_access_type:
+                                grants = DataAccessGrant.objects.filter(application=self)
+                                for grant in grants:
+                                    grant.update_expiration_date()
+                                end_time = time.time()
+                                update_stats = {
+                                    "elapsed_seconds": end_time - start_time,
+                                    "number_of_grants_updated": grants.count(),
+                                    "grant_update_complete": datetime.now().strftime("%m/%d/%Y, %H:%M:%S")
+                                }
+                                log_dict.update(update_stats)
                             app_type_changed = True
-                            end_time = time.time()
-                            delete_stats = {
-                                "elapsed_seconds": end_time - start_time,
-                                "number_of_grant_deleted": dag_deleted[0],
-                                "grant_delete_complete": datetime.now().strftime("%m/%d/%Y, %H:%M:%S")
-                            }
-                            log_dict.update(delete_stats)
                             logger.info(log_dict)
                 except Application.DoesNotExist:
                     # new app
                     pass
-                self.copy_client_secret()
-                super().save(*args, **kwargs)
                 if app_type_changed:
-                    log_dict.update({"application_saved_and_grants_deleted": "Yes"})
+                    log_dict.update({"application_saved_and_grants_updated_or_deleted": "Yes"})
                     logger.info(log_dict)
         else:
             self.copy_client_secret()
